# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DepartmentHub - Academic Event & Progress Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error

  # Events endpoints
...
  /ai/summaries:
    get:
      operationId: listAiSummaries
      tags: [ai]
      summary: List AI-generated summaries
      parameters:
        - name: departmentId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of AI summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiSummary"
    post:
      operationId: generateAiSummary
      tags: [ai]
      summary: Generate a new AI summary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiSummaryInput"
      responses:
        "201":
          description: AI summary generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSummary"
  /ai/summaries/{id}:
    get:
      operationId: getAiSummary
      tags: [ai]
      summary: Get a specific AI summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSummary"
...
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List all departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a department (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Department created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{id}:
    get:
      operationId: getDepartment
      tags: [departments]
      summary: Get a department by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Department found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    patch:
      operationId: updateDepartment
      tags: [departments]
      summary: Update a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentUpdate"
      responses:
        "200":
          description: Department updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RegisterInput:
      type: object
      required: [email, password, firstName, lastName, role]
      properties:
        email:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [ADMIN, FACULTY, STUDENT, MANAGEMENT]
        departmentId:
          type: ["integer", "null"]
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    # User schemas
    User:
      type: object
      required: [id, email, firstName, lastName, role, isActive, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [ADMIN, FACULTY, STUDENT, MANAGEMENT]
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        isActive:
          type: boolean
        lastLogin:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
...
    Department:
      type: object
      required: [id, name, code, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        headUserId:
          type: ["integer", "null"]
        headUserName:
          type: ["string", "null"]
        memberCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    DepartmentInput:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
        code:
          type: string
        headUserId:
          type: ["integer", "null"]
    DepartmentUpdate:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        headUserId:
          type: ["integer", "null"]

    # Event schemas
...
    AiSummary:
      type: object
      required: [id, type, content, generatedAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [WEEKLY, MONTHLY, PROGRESS, TREND]
        content:
          type: string
        insights:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        dateRangeStart:
          type: ["string", "null"]
        dateRangeEnd:
          type: ["string", "null"]
        generatedAt:
          type: string
    AiSummaryInput:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [WEEKLY, MONTHLY, PROGRESS, TREND]
        departmentId:
          type: ["integer", "null"]
        dateRangeStart:
          type: string
        dateRangeEnd:
          type: string
...