# lib/api-zod/src/generated/types/aiSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DepartmentHub - Academic Event & Progress Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { AiSummaryType } from './aiSummaryType';

export interface AiSummary {
  id: number;
  type: AiSummaryType;
  content: string;
  /** @nullable */
  insights?: string | null;
  /** @nullable */
  departmentId?: number | null;
  /** @nullable */
  departmentName?: string | null;
  /** @nullable */
  dateRangeStart?: string | null;
  /** @nullable */
  dateRangeEnd?: string | null;
  generatedAt: string;
}