# lib/api-zod/src/generated/types/aiSummaryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DepartmentHub - Academic Event & Progress Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { AiSummaryInputType } from './aiSummaryInputType';

export interface AiSummaryInput {
  type: AiSummaryInputType;
  /** @nullable */
  departmentId?: number | null;
  dateRangeStart?: string;
  dateRangeEnd?: string;
}